pyproject.toml
[tool.poetry]
name = "telegram-ai-agent"
version = "0.1.0"
description = "A Python library for managing Telegram AI agents with outbound and inbound messaging capabilities"
authors = ["Ivan Fokeev <vanya@vanya.cc>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
telethon = "^1.37.0"
pysocks = "^1.7.1"
phidata = "^2.4.37"
openai = "^1.49.0"
langchain = "^0.3.1"
langchain-experimental = "^0.3.2"
tiktoken = "^0.7.0"
httpx = "^0.27.2"
python-dotenv = "^1.0.1"
streamlit = "^1.38.0"
pandas = "^2.2.3"
sqlalchemy = "^2.0.35"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "24.10.0"
isort = "5.13.2"
pre-commit = "4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
